code: Don't use try-catch as function definition
This breaks MSVC and results in leaked exceptions.

+++ source/nvidia/ar/nvidia-ar.cpp
@@ -90,7 +90,7 @@ streamfx::nvidia::ar::ar::ar() : _library(), _model_path()
 		DWORD env_size = GetEnvironmentVariableW(L"NVAR_MODEL_PATH", nullptr, 0);
 		if (env_size > 0) {
 			std::vector<wchar_t> buffer(static_cast<size_t>(env_size) + 1, 0);
-			env_size    = GetEnvironmentVariableW(L"NVAR_MODEL_PATH", buffer.data(), buffer.size());
+			env_size    = GetEnvironmentVariableW(L"NVAR_MODEL_PATH", buffer.data(), static_cast<DWORD>(buffer.size()));
 			_model_path = std::wstring(buffer.data(), buffer.size());
 
 			// The SDK is location one directory "up" from the model path.
